Use Hiera 2.x backend API
Before this commit the rspec backend still used the deprecated Hiera 1.x
backend api that interprets nil values returned by #lookup as a lookup
failure (i.e. it was not possible to use keys with nil values).

The readme for the backend specifically states that it is meant to
interface with Hiera version 3.2.0 so it should not be using 1.x.

This commit changes the arity and semantics of the lookup function to
match that of the 2.x backend API.